+++ server/go.mod
@@ -8,6 +8,7 @@ require (
 	github.com/alecthomas/participle/v2 v2.0.0-alpha8
 	github.com/brianvoe/gofakeit/v6 v6.17.0
 	github.com/denisbrodbeck/machineid v1.0.1
+	github.com/fluidtruck/deepcopy v1.0.0
 	github.com/fullstorydev/grpcurl v1.8.6
 	github.com/gogo/protobuf v1.3.2
 	github.com/golang-migrate/migrate/v4 v4.15.2
@@ -54,7 +55,6 @@ require (
 	github.com/docker/go-connections v0.4.0 // indirect
 	github.com/docker/go-units v0.4.0 // indirect
 	github.com/felixge/httpsnoop v1.0.2 // indirect
-	github.com/fluidtruck/deepcopy v1.0.0 // indirect
 	github.com/go-logr/logr v1.2.3 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
 	github.com/go-redis/redis/v7 v7.4.1 // indirect
