+++ server/testdb/postgres/transactions_test.go
@@ -1,4 +1,4 @@
-package testdb_test
+package postgres_test
 
 import (
 	"context"
