--- server/testdb/results.go
@@ -1,1 +1,0 @@
-package testdb
