+++ server/testdb/testdb.go
@@ -1,0 +1,12 @@
+package testdb
+
+import (
+	"errors"
+
+	"github.com/kubeshop/tracetest/server/id"
+)
+
+var (
+	IDGen       = id.NewRandGenerator()
+	ErrNotFound = errors.New("record not found")
+)
